This is the first version of the initial revision 1a1 development tree.  This includes new source files and solution files supporting compilation and linking using Miscosoft Visual C++ 2008 for initial portation and experimentation with Microsoft Windows running as a console application.
Note that at the time of this checkin, the Microsoft windows version of the ptpv2d code does compile and link without errors, but it is at this time only paritally functional.  I wanted to go ahead and get it published to see if we could get others interested in PTP on Windows with more experience.

Also this version has now been cleaned up to make it so that it also does now compile under standard PC based linux.  I've tried to do my best to clean out some of the MPC831X specific code such that it should run OK on a standard Linux PC as well as still supporting Hardware based timestamping on the Freescale MPC831X family.

This is a development version so myself and other members of the ptpv2d team will be doing dynamic updates of this code as we see fit.  If the latest version of the development tree is not what you want, you should be able to retrieve older versions using SVN.

Email all comments, questions and questions to ptpv2d@googlegroups.com

For bug reports, please log them under the Google code groups at: http://code.qoogle.com/p/ptpv2d/issues/list and select "New issue".

Reminder: This code base is copyrighted and licensed under the GPL Version 2 license.  Any other use of this code base is not permitted (in other words, this is open source and it is to stay that way, if you use this source, you must abide by GPL Version 2 license requirements, no exceptions).